Organize files and folders in repo
